requirements.txt
Django==3.1.5
incremental==17.5.0
django-cors-headers==3.3.0
gunicorn==20.0.4
django-heroku
djangorestframework
django-cors-headers==3.3.0
daphne==3.0.1
gunicorn
channels==3.0.3
cryptography==3.3.1